+++ backend/src/main/java/com/techeer/backend/api/feedback/converter/FeedbackConverter.java
@@ -1,11 +1,12 @@
-package com.techeer.backend.api.feedback.domain;
+package com.techeer.backend.api.feedback.converter;
 
+import com.techeer.backend.api.feedback.domain.Feedback;
 import com.techeer.backend.api.feedback.dto.FeedbackResponse;
 import com.techeer.backend.api.resume.domain.Resume;
 
 public class FeedbackConverter {
 
-	public static FeedbackResponse of(Resume resume, Feedback feedback) {
+	public static FeedbackResponse toFeedbackResponse(Resume resume, Feedback feedback) {
 		return FeedbackResponse.builder()
 			.feedbackId(feedback.getId())
 			.resumeId(resume.getId())
